Leftover cast (Fix #4416)

+++ radio/src/gui/128x64/model_logical_switches.cpp
@@ -115,7 +115,7 @@ void menuModelLogicalSwitchOne(event_t event)
           v1_max = 122;
         }
         else {
-          v1_val = (uint8_t)cs->v1;
+          v1_val = cs->v1;
           drawSource(CSWONE_2ND_COLUMN, y, v1_val, attr);
           INCDEC_SET_FLAG(EE_MODEL | INCDEC_SOURCE);
           INCDEC_ENABLE_CHECK(isSourceAvailable);
